requirements.txt
dnspython
click
influxdb
